worker/requirements.txt
redis==5.2.1
python-dotenv==1.0.1